Added feature for transmitting roadmap selected blocks to a new auxiliary leadsheet window.
If no blocks are selected, then all blocks will be selected.

If the roadmap is empty, nothing happens.

Right now, a new leadsheet is opened every time this action is taken. We might well want to change this to add to the current auxiliary window rather than create a new one.