Related to the previous commit. Now the first time chords are transferred from a roadmap to a new leadsheet. On subsequent times, they are added to the previous leadsheet.
I am noting that the last chord of a selection is intermittently omitted in the transfer. There may be something wrong with the underlying method.

Also, if the leadsheet window is deleted by the user, then a new attempt is made to transfer, there is likely going to be a problem. This needs to be fixed.