Cosmetic changes to widgets at top of RoadMapFrame. Added Titled Border to feature width slider.
In order to get the scroll bars on the roadmap panel to work, I think the RoadMapPanel itself will have to be large enough so that it can't be seen at one time in the ScrollPane. This might not be the case, or if it is, then maybe a panel has to be added in the GUI builder.