Now playing from the Road Map plays in the first style in the leadsheet from which the Road Map was opened.
Currently only the first style is used. If there are multiple sections with different styles, the styles of other than the first section will not be used.

We may eventually consider playing with styles that are specified on a section-by-section basis.

Currently the original chords-only playback can be obtained by having the original leadsheet set to no-style. We may eventually provide style selections from the road map itself.