Added text field to RoadMapFrame for entering an arbitrary sequence of chords in Leadsheet notation.
If melody is also entered, a warning will be issued, as they make no sense.

Currently the ChordPart thus created is reparsed into strings for analysis. Eventually, it would be better to use the Chords  in the part directly.

The message from this warning popped up repeatedly beofre and after this change. I've commented it out for now. Please restore it for any meaningful cases.

           //ErrorLog.log(ErrorLog.WARNING, "Cannot find joins in an ArrayList "
            //        + "of size 0");