Added new constructor to brickdictionary.ChordBlock so that a ChordBlock can be constructed directly from a data.Chord without going to a String and passing that to a constructor.
Chords in the roadmap, either from an original Leadsheet or by being typed into the textual entry will be of this form.

An advantage is that members of Chord are pre-validated to be understood by other parts of Impro-Visor.

Still to be added is a complementary constructor of a Chord from a ChordBlock.