Added to class polya.Arith static method
int long2int(long value)

for converting longs, such as returned from the Polya Tokenizer, to ints.

This should only be used if it is perceived to be safe, as in dealing with durations in terms of number of slots.

It will throw an Error if the value is out of range

Example in data.Chord:

    rhythmValue = polya.Arith.long2int(block.getDuration());

Please use this rather than open-coding equivalent logic.