Made several changes for interactions between leadsheet (Notate) and RoadMap:
Now there is a RoadMap menu in Notate, instead of the use of the Utilities menu.

Now there is a Leadsheet menu in RoadMapFrame with two options: Append to leadsheet and open new leadsheet, then append.

The "To Leadsheet" button is gone.

Other buttons are rearranged, to reflect common sequence of use.

Automatic playing of the leadsheet no longer happens.