This is just a checkpoint.
Despite my efforts to fix the problem, it still exists:

When a new leadsheet is opened in the same window, the metre (time signature) of the roadmap does not get updated.

I don't see clearly where the metre is used in drawing the roadmap. It might help if I did.

Barring a simple solution, I will try deleting the original roadmap and creating a new one.