Workaround, I hope temporary, to prevent losing measures/line formatting when launching with a most recent file.
In order to get the formatting, the leadsheet is repopulated. This shows up on big leadsheets, such as _test.ls

It seems to have something to do with whether the Notate window is already open when the file loads. If it is open, the syndrome does not occur.

An alternative is to sacrifice the formatting and reopen the leadsheet, but this is not good either.