Starting version 5.07.
Solved a problem on Linux and Windows wherein vertical scrollbar on Brick Dictionary disappears after expanding a node.

private void libraryTreeTreeExpanded(javax.swing.event.TreeExpansionEvent evt) {                                         
        libraryTree.setPreferredSize(
                new Dimension(libraryTree.getPreferredSize().width,
                libraryTree.getRowCount() * libraryTree.getRowHeight()));
    }     

Under non-Mac systems, changing the JTree size in this way destroys the vertical scrollbar in the surrounding JScrollPane. It would seem to be a bug in those implementations.

For now, I'll just have to make the JTree a fixed (large) size so that it works on all platforms.