Checkpoint. Restored non-grammar behavior in lickgen.
Ran in to a snag with an array index out of range when Fill Melody is pressed the first time. 

Sorry for this horrible hack in LickGen.java, but have to recover.

The problem may be due to switching from Vector to ArrayList, back in rev. 721, but unable to pin it down so far.

                // Temporary HACK ALERT; Preventing index out of range. 
                // Moving to rev. 790
                // May have something to do with transition from Vector to ArrayList
                if( section >= probs.size() )
                  {
                  section = probs.size()-1;
                  }