Huge instrumentation debugging foray to try to find out why drum pattern extraction was not working.
It doesn't seem like this code could ever have worked as it stood (Jim and Brandy). There are too many erroneous casts (using array element 0, which is a string different than the elements.

Maybe it's the result of some commit mess up, I don't know.

Anyway, it seems to be working now, and I am committing prior to taking out the instrumentation.