Now currentSlot in Grammar is set to the current slot at the time generateRhythmFromGrammar is called. The latter now takes startSlot as an argument for lick generation.
However, currently startSlot is only set statically; it is not changed during playing. This aspect of dynamic trading is yet to be worked out.

The functioning of this scheme can be tested by improvising on a selection that does not start at the beginning of the chorus. It has been tested in this fashion.