Grammar "splice" operator added on BOB branch.
 * Evaluates a special form (splice <operator> <arg> ...)
 * (<operator> <arg> ...) must return a list.
 * Currently the only operator available is literal, e.g.
 * a rule RHS of the form
 * (C4 (splice literal C8 C8) (splice literal C16 C16 C16 C16) R4)
 * returns (C4 C8 C8 C16 C16 C16 C16 R4)
 * 
 * It is expected that additional splice-oriented operators will be added,
 * and not all will just use literal arguments.

Also adds a temporary hack in the Bass Pattern area. If no pattern is found to fit a desired number of slots, the first bass pattern is used, and that pattern will extend beyond the duration of the space into which the pattern fits. This is a workaround until a more robust solution can be found to the problem of making one pattern extend over more than one chord in a sequence. The intention is that this device is used for styles containing only one bass pattern.