Checkpoint on JACK branch.
At present, when a new chorus is added by the user, a new chord part is created by copying the current chord part. A name for the new part is generated CP_1, CP_2, etc. (If a name is already in use, that name will be skipped over.)

Modifications to the chords and length of the new parts do not affect the old ones. 

However, length modifications incorrectly do not affect the playback, so this needs work.

The variable chordProg in Score no longer exists. Use of it are replaced with getDefaultChordPart().

An assessment needs to be done of the playability of various temporary Scores, such as are used in the StyleEditor, etc.