Checkpoint on JACK branch.
Now chord parts and tabs of various sizes are mostly working and the rendered MIDI plays as desired.

Still to be addressed is the way the Notate window updates on playing. Because a fixed chord part size was formerly assumed, the play tracking line and tab change no longer work.