Checkpoint. Integrating code from branch BOB into HEAD is nearing completion.
Now auto-improvisation on play functionality is available in the HEAD branch.
There are several Auto-Improv menu items. The default is no auto-improvisation. This enables the Improvise button for use.
In other modes, Impro-Visor will trade with the user upon Play.
The Improvise button should not be used for trading. If it is used, it will cause a return to the default mode.

Several modifications are needed:
(1) The code from David's committed revision 1235 is not present. I, or David, will work on this. As soon as it is done, development on the BOB branch should be abandoned.

(2) The Lickgen window opens when auto-improvisation is used, but is immediately hidden. This behavior should be cleaned up, and having LickGenSettings class may help, as the reason for it is to initialize the LickGen parameters.

(3) On auto-improvisation the melody instruments get confused. This will be fixed by setting the instruments appropriately.

(4) Sometimes a trading melody is not inserted when it should be. This will be investigated and hopefully remedied. I think it has to do with not hitting the slot number exactly.