Now the new synchronization & expectancy parameters are set from the grammar file, if present. However, only the useSyncopation value is currently used. I will leave it up to David Halpern to connect uses of the other parameters, all in Lickgen, to their appropriate position. If not present, the parameters are set to default values, as follows:
private static final boolean defaultUseSyncopation      = false;
private static final String defaultSyncopationType      = "C";
private static final double defaultSyncopationValue     = 0.5;
private static final double defaultExpectancyMultiplier = 0.5;
private static final double defaultExpectancyConstant   = 0.1;

These values can be modified in the Grammar Editor, and if the grammar is saved, the modified values will be saved along with it.

An intermittent error having to do with a pitch value of 1009 is manifest in MelodyPart.java. When hit, this results in the Notate window blowing up. I put in protective code for the time being, so that it shows as
out-of-range pitch 1009 in makeAccidentals
rather than blowing up. However, as 1009 is related to expectancy, I am hoping that David will be able to chase this down.