Checkpoint on linking AudioSettings to corresponding GUI elements. Now widget values are determined from the defaults and setters in AudioSettings. AudioSettings are not currently settable from the GUI. This needs to be added.
Note that Slider scale values must be integers, so these are 1000, 10, and 100 times the actual value of the respective parameters.
Note that the Note Resolution needs to be expressed in slots, not number of sub-divisions, at present, and that the resolution model is statically shared with all other uses of this model. We may want to change this.