Checkpoint. Added a menu in Chorus Preferences for Aux Instrument, to be used for the improvising instrument.
An obstacle to making this function was that the a program change event has to be sent on each note of the original melody, to prevent the improving melody from usurping the channel for its instrument. A future solution might use a different channel for the melody instrument. This change is in MelodyPart.java.

Also I'm noticing a problem wherein the first time or two a chorus is played with trading, on, say, the chord grammar, the generated melody is just some very long tones, around B on the staff. I tried to trace this back, and thought it started around revision 1274 or 1275, but I don't have a definitive revision number. The problem seems nondeterministic and goes away after the first couple of plays. Further investigation is needed.

Refactored NonExistentParameterException to go in /src/imp/util rather than /src/imp/lickgen, as it is being used for general parameters as well as lickgen parameters.