Preliminary version of roadmap saving and re-loading. Must be run on leadsheets where there is no old version of saved roadmap.
Now a roadmap is read from the leadsheet file. However, nothing is done with it except possibly printing it for reference.
Uncommenting the indicated lines in ChordPart.java will show the difference.
It seems reasonably close except for the keymap, which is not being reconstructed properly, due to that part of the code not having been initialized at the time the roadmap is read. 
(The BrickDictionary has to be read first.)

Once the above issues are worked out, the reconstructed roadmap should be usable as *the* roadmap.

In ChordPart.java:
    public void setRoadmapPoly(Polylist roadmapPoly)
    {
        this.roadmapPoly = roadmapPoly;
        //System.out.println("The original roadmap poly is " + Formatting.prettyFormat(roadmapPoly));
        roadmap = new RoadMap(roadmapPoly);
        //System.out.println("The reconstructed roadmap is " + Formatting.prettyFormat(roadmap.toPolylist()));
    }