Adding method to Notate:
public ArrayList<Block> getRoadMapBlocks()

so that learning based on bricks can progress. 

Note however, that depending on when this method is called, a race condition in cykparser may be exposed. This should be fixed. Example (there can be more than one such problem. It is timing-dependent):

Exception in thread "main" java.util.ConcurrentModificationException
	at java.util.LinkedList$ListItr.checkForComodification(LinkedList.java:761)
	at java.util.LinkedList$ListItr.next(LinkedList.java:696)
	at imp.cykparser.CYKParser.findTerminal(CYKParser.java:486)
	at imp.cykparser.CYKParser.fillTable(CYKParser.java:351)
	at imp.cykparser.CYKParser.parse(CYKParser.java:669)
	at imp.roadmap.RoadMapFrame.analyze(RoadMapFrame.java:3869)
	at imp.roadmap.RoadMapFrame.getFullRoadMap(RoadMapFrame.java:4653)
	at imp.gui.Notate.getRoadMapBlocks(Notate.java:26193)
	at imp.ImproVisor.<init>(ImproVisor.java:390)
	at imp.ImproVisor.main(ImproVisor.java:492)