Initial commit of grammar capability for conditioning on brick presence.
Now the probability component (4th element) of a grammar rule can invoke an evaluable expression

	(builtin brick ...name-of-brick...)

which will return 1.0 if a brick of that type begins at the current slot during grammar expansion.
Otherwise the expression returns 0.0. So the production is conditionally applied if that brick is present.

ABricktest01.grammar contains an example with 4 brick types, 3 of which are 2-bar and one of which is 4-bar.
These bricks occur in the tune BlueMoon, a leadsheet for which can be found in leadsheets/insights/BlueMoon.ls

NOTE: The roadmap must have already been computed for the tune in order to use this. If it is not,
then the probabilities will evaluate to 0. Use manual evaluation or evaluation on load menu item.

There are still some things to be worked out with how the probabilities work, not so much with this method
but with probabilities in general.

Currently debug output is left on. It can be commented out in Grammar.java. Example output for BlueMoon is

At slot 0 using brick POT
At slot 960 using brick POT
At slot 1920 using brick POT
At slot 2880 using brick POT
At slot 3840 using brick POT
At slot 4800 using brick POT
At slot 7680 using brick 7-Chord-Dropback
At slot 9600 using brick Straight-Cadence
At slot 10560 using brick Dogleg-Slow-Launcher