Now a grammar using brick productions will fill the entire chorus, if it fills at all. Some fraction of the time, no beats are generated, with or without bricks. This was always true to some extent, although it seems more common now.
I'm noticing that, even when an abstract or relative pitch melody is generated (as evidenced in the lickgen frame), the melody does not always get filled. Therefore there is a problem with fillMelody, although it doesn't seem be related to Grammar. I will investigate further.

It is also no longer necessary to specify a list of terminals in the grammar file, as terminals are now determined by internal parsing, using a method isAbsractNote which in turn uses a new method isDuration in Duration.java.