Now improvisation works over both standard and brick grammars. This is accomplished by modifying the mode in which productions are employed. Formerly expansion was driven by a single starting P goal, that is a non-terminal of the form (P n) where n is a number of slots to be filled, (P is the start symbol of the grammar.)
Now instead, expansion is driven by the number of slots to be filled, and multiple P goals may be generated in sequence. As the productions expand, we keep track of the length of the melody that will ultimately be generated (by measuring the duration of the abstract or relative pitch melody). If the generated melody is not long enough, another P goal is introduced to try to generate the residual requirement. This continues until the overall requirement is met or exceeded. 

At the end, the generated abstract melody is truncated to be no longer than the desired length. This may leave a few beats unfilled at the end, but this is not regarded as a problem.

This version of Grammar.java is a rough draft, and clean up will follow.