Added melodyPart2Relative to NoteConverter. This will take an arbitrary melody part with a chord part and convert it into an abstract relative-pitch melody, which can then be "filled" over any chords.
In ThemeWeaver lines 1177-1189 I placed code that illustrates this by converting the melody extracted from the leadsheet into abstract, then filling it. This code is currently commented-out, but can be uncommented to illustrate.

I think that, in the long run, using the relative-pitch melody will be superior to using the absolute melody. For one thing, it is automatically transposable. For another, the approximate shape of the original is maintained.