Nathan Kim installed Feature Request #80:
In the main window there is a pair of arrow buttons used for stepping through chords. (They are colored green.) It would be nice if there were a remote version of the same buttons in the voicing keyboard, so that the voicings could be viewed without returning to the main window each time.